refactor(settings): extract Networks enable-wizard; NetworksComponent 1261→692 lines (PR-U3 slice 9)
Third and final dialog extraction: pull the 3-step Enable-Networks wizard out of NetworksComponent into
NetworkEnableWizardComponent. It owns the whole flow — hostname + options, the generated Cloudflare-tunnel
commands (Windows/Linux), and the local-agent connector probe→bootstrap→execute path — and emits
`enabled(url)` when this brain gets a public URL; the host adopts it and clears needsNetworkEnable.
detectLocalOs moved with it; isLocalOrPrivateUrl stays (ngOnInit still derives needsNetworkEnable). Also
removed the now-dead dialog/wizard CSS and the unused ModalDirective import from the parent.

Behavior unchanged. The 10 enable-wizard characterization tests moved to
network-enable-wizard.component.spec.ts (success now asserts the `enabled`/`close` outputs); the parent
keeps an onEnabled test. 39 networks specs across the 4 files pass; AOT build clean.

With all three dialogs extracted, NetworksComponent is down from 1261 to 692 lines (−45%) — the
settings-UX audit's "worst offender" is now maintainable. Children: create 200, join 273, enable 355.
